chore: prefixed split minarch files with ma_ instead
