+++ workspace/all/minarch/ma_input.h
@@ -7,5 +7,5 @@ void input_poll_callback(void);
 int16_t input_state_callback(unsigned port, unsigned device, unsigned index, unsigned id);
 void Input_init(const struct retro_input_descriptor *vars);
 
-// Menu functions (defined in minarch_menu.c, see minarch_menu.h)
-#include "minarch_menu.h"
+// Menu functions (defined in ma_menu.c, see ma_menu.h)
+#include "ma_menu.h"
