+++ workspace/all/minarch/makefile
@@ -22,8 +22,8 @@ SDL?=SDL
 TARGET = minarch
 PRODUCT= build/$(PLATFORM)/$(TARGET).elf
 INCDIR = -I. -I./libretro-common/include/ -I../common/ -I../../$(PLATFORM)/platform/
-SOURCE = $(TARGET).c minarch_cheats.c minarch_rewind.c minarch_audio.c minarch_input.c \
-         minarch_options.c minarch_frontend_opts.c minarch_saves.c minarch_video.c minarch_core.c minarch_game.c minarch_environment.c minarch_config.c minarch_menu.c minarch_runframe.c \
+SOURCE = $(TARGET).c ma_cheats.c ma_rewind.c ma_audio.c ma_input.c \
+         ma_options.c ma_frontend_opts.c ma_saves.c ma_video.c ma_core.c ma_game.c ma_environment.c ma_config.c ma_menu.c ma_runframe.c \
          ../common/scaler.c ../common/utils.c ../common/config.c ../common/api.c \
          ../common/notification.c ../../$(PLATFORM)/platform/platform.c
