+++ workspace/all/minarch/minarch.c
@@ -6,19 +6,19 @@
 #include "notification.h"
 #include "ra_integration.h"
 
-#include "minarch_internal.h"
-#include "minarch_cheats.h"
-#include "minarch_audio.h"
-#include "minarch_input.h"
-#include "minarch_options.h"
-#include "minarch_frontend_opts.h"
-#include "minarch_saves.h"
-#include "minarch_video.h"
-#include "minarch_core.h"
-#include "minarch_game.h"
-#include "minarch_environment.h"
-#include "minarch_config.h"
-#include "minarch_runframe.h"
+#include "ma_internal.h"
+#include "ma_cheats.h"
+#include "ma_audio.h"
+#include "ma_input.h"
+#include "ma_options.h"
+#include "ma_frontend_opts.h"
+#include "ma_saves.h"
+#include "ma_video.h"
+#include "ma_core.h"
+#include "ma_game.h"
+#include "ma_environment.h"
+#include "ma_config.h"
+#include "ma_runframe.h"
 
 ///////////////////////////////////////
